简单的完善修改
git-svn-id: https://10.211.55.3/svn/chelib@282 5bab769a-80ee-f445-9178-c608493bda25

+++ trunk/src/che_bitmap.cpp
@@ -198,10 +198,10 @@ HE_BOOL		CHE_Palette::GetNearColorIndex( HE_ARGB color, HE_DWORD & indexRet ) co
 	{
 		srcColor = m_pPalette[i];
 		
-		tempLessCY = abs( ( color & 0x000000FF ) - ( srcColor & 0x000000FF )  );
-		tempLessCY += abs( ( color >> 8 & 0x000000FF ) - ( srcColor >> 8 & 0x000000FF ) );
-		tempLessCY += abs( ( color >> 16 & 0x000000FF ) - ( srcColor >> 16 & 0x000000FF ) );
-		tempLessCY += abs( ( color >> 24 & 0x000000FF ) - ( srcColor >> 24 & 0x000000FF ) );
+		tempLessCY =  abs( (long)( ( color & 0x000000FF ) - ( srcColor & 0x000000FF ) ) );
+		tempLessCY += abs( (long)( ( color >> 8 & 0x000000FF ) - ( srcColor >> 8 & 0x000000FF ) ) );
+		tempLessCY += abs( (long)( ( color >> 16 & 0x000000FF ) - ( srcColor >> 16 & 0x000000FF ) ) );
+		tempLessCY += abs( (long)( ( color >> 24 & 0x000000FF ) - ( srcColor >> 24 & 0x000000FF ) ) );
 		if ( tempLessCY < lessCY )
 		{
 			lessCY = tempLessCY;
